Fix typeInfo ordering
Unfortunately with the following visitor

```
{
  '*'()           { console.log('*:enter') },
  '*:exit'()      { console.log('*:exit') },
  'Field'()       { console.log('Field:enter') },
  'Field:exit'()  { console.log('Field:exit') },
}
```

Eslint gets us

```
*:enter
Field:enter
*:leave
Field:leave
```

Instead of the more sensible

```
*:enter
Field:enter
Field:leave
*:leave
```

This means that neither the general node nor specific node can be relied
on to bracket the other kind: they will be out of order either on the
way down the tree or on the way back up.

However, GraphQL's `TypeInfo` expects to enter before its rules enter
and to leave before its rules leave.  Because of our failure to bracket
the rule visitor correctly due to ESLint's odd traversal, type checking
would have the wrong field types in certain cases, as with the
following:

```
{
  cat {
    affection(bittenRecently: false) {
      value
    }
  }
}
```

When leaving the `value` field, if we leave the typeInfo first, `value`
will be out of step with the field definition stack and think that it
needs to check `bittenRecently` as a required parameter.

This commit fixes the traversal issue by executing everything within the
generic `*` nodes in the ESLint visitor so that typeInfo can bracket the
rule visit.